+++ docs/contents/changelog.rst
@@ -2,6 +2,18 @@ Changelog
 =====================================
 The source code is hosted and maintained on `github <https://github.com/dschenck/easytree/>`_.
 
+Version 1.0.0 (2024-11-12)
+--------------------------
+    - removed deprecated :code:`easytree.Tree` class 
+    - removed deprecated :code:`easytree.Node` class alias 
+    - removed deprecated :code:`easytree.new` function
+    - removed deprecated :code:`easytree.load` function
+    - removed deprecated :code:`easytree.loads` function
+    - removed deprecated :code:`easytree.dump` function
+    - removed deprecated :code:`easytree.dumps` function
+    - removed deprecated :code:`easytree.serialize` function
+    - fixed bug on :code:`easytree.undefined.get`
+
 Version 0.2.4 (2023-11-25)
 --------------------------
     - fixed a bug where :code:`dict.setdefault` would never set the default value